needs to be BigDecimal

+++ src/main/java/com/auction/cart/dto/ProductRequest.java
@@ -5,11 +5,9 @@ import java.math.BigDecimal;
 
 public record ProductRequest(
         Long productId,
-
-       String name,
-      // String description,
-
-       double price
+        String name,
+        // String description,
+        BigDecimal price
 
 ) {
